splitted out file io
