+++ NanoVNASaver/Windows/Files.py
@@ -19,6 +19,8 @@
 import logging
 
 from PyQt5 import QtWidgets, QtCore
+from NanoVNASaver.Touchstone import Touchstone
+from NanoVNASaver.RFTools import Datapoint
 
 logger = logging.getLogger(__name__)
 
@@ -39,9 +41,9 @@ class FilesWindow(QtWidgets.QWidget):
         load_file_control_layout = QtWidgets.QFormLayout(load_file_control_box)
 
         btn_load_sweep = QtWidgets.QPushButton("Load as sweep")
-        btn_load_sweep.clicked.connect(self.app.loadSweepFile)
+        btn_load_sweep.clicked.connect(self.loadSweepFile)
         btn_load_reference = QtWidgets.QPushButton("Load reference")
-        btn_load_reference.clicked.connect(self.app.loadReferenceFile)
+        btn_load_reference.clicked.connect(self.loadReferenceFile)
         load_file_control_layout.addRow(btn_load_sweep)
         load_file_control_layout.addRow(btn_load_reference)
 
@@ -52,11 +54,11 @@ class FilesWindow(QtWidgets.QWidget):
         save_file_control_layout = QtWidgets.QFormLayout(save_file_control_box)
 
         btn_export_file = QtWidgets.QPushButton("Save 1-Port file (S1P)")
-        btn_export_file.clicked.connect(lambda: self.app.exportFile(1))
+        btn_export_file.clicked.connect(lambda: self.exportFile(1))
         save_file_control_layout.addRow(btn_export_file)
 
         btn_export_file = QtWidgets.QPushButton("Save 2-Port file (S2P)")
-        btn_export_file.clicked.connect(lambda: self.app.exportFile(4))
+        btn_export_file.clicked.connect(lambda: self.exportFile(4))
         save_file_control_layout.addRow(btn_export_file)
 
         file_window_layout.addWidget(save_file_control_box)
@@ -64,3 +66,64 @@ class FilesWindow(QtWidgets.QWidget):
         btn_open_file_window = QtWidgets.QPushButton("Files ...")
         btn_open_file_window.clicked.connect(
             lambda: self.app.display_window("file"))
+
+    def exportFile(self, nr_params: int = 1):
+        if len(self.app.data11) == 0:
+            QtWidgets.QMessageBox.warning(
+                self, "No data to save", "There is no data to save.")
+            return
+        if nr_params > 2 and len(self.app.data21) == 0:
+            QtWidgets.QMessageBox.warning(
+                self, "No S21 data to save", "There is no S21 data to save.")
+            return
+
+        filedialog = QtWidgets.QFileDialog(self)
+        if nr_params == 1:
+            filedialog.setDefaultSuffix("s1p")
+            filedialog.setNameFilter(
+                "Touchstone 1-Port Files (*.s1p);;All files (*.*)")
+        else:
+            filedialog.setDefaultSuffix("s2p")
+            filedialog.setNameFilter(
+                "Touchstone 2-Port Files (*.s2p);;All files (*.*)")
+        filedialog.setAcceptMode(QtWidgets.QFileDialog.AcceptSave)
+        selected = filedialog.exec()
+        if not selected:
+            return
+        filename = filedialog.selectedFiles()[0]
+        if filename == "":
+            logger.debug("No file name selected.")
+            return
+
+        ts = Touchstone(filename)
+        ts.sdata[0] = self.app.data11
+        if nr_params > 1:
+            ts.sdata[1] = self.app.data21
+            for dp in self.app.data11:
+                ts.sdata[2].append(Datapoint(dp.freq, 0, 0))
+                ts.sdata[3].append(Datapoint(dp.freq, 0, 0))
+        try:
+            ts.save(nr_params)
+        except IOError as e:
+            logger.exception("Error during file export: %s", e)
+            return
+
+    def loadReferenceFile(self):
+        filename, _ = QtWidgets.QFileDialog.getOpenFileName(
+            filter="Touchstone Files (*.s1p *.s2p);;All files (*.*)")
+        if filename != "":
+            self.app.resetReference()
+            t = Touchstone(filename)
+            t.load()
+            self.app.setReference(t.s11data, t.s21data, filename)
+
+    def loadSweepFile(self):
+        filename, _ = QtWidgets.QFileDialog.getOpenFileName(
+            filter="Touchstone Files (*.s1p *.s2p);;All files (*.*)")
+        if filename != "":
+            self.app.data11 = []
+            self.app.data21 = []
+            t = Touchstone(filename)
+            t.load()
+            self.app.saveData(t.s11data, t.s21data, filename)
+            self.app.dataUpdated()
